+++ src/entry-cjs.js
@@ -1,5 +1,7 @@
 const style = require('./style')
 const nostyle = require('./nostyle')
+const saveAs = require('./saveAs')
 
 module.exports = style
 exports.nostyle = nostyle
+exports.saveAs = saveAs