+++ src/entry-esm.js
@@ -1,7 +1,9 @@
 import style from './style.modern'
 import nostyle from './nostyle.modern'
+import saveAs from './saveAs.modern'
 
 export default style
 export {
-  nostyle
+  nostyle,
+  saveAs
 }