+++ src/style.js
@@ -1,27 +1,8 @@
 import { write } from 'xlsx-style-vite'
-import { getSheets, s2ab, saveAs } from './utils'
-import { getCellStyle } from './utils.style'
+import download from './utils'
+import getCellStyle from './utils/cellStyle'
 
-export default ({
-  sheets,
-  filename = 'excel.xlsx',
-  hiddenHeader = false,     // 导出表格时，是否隐藏表头。默认显示表头
-}) => {
-  const Sheets = getSheets({
-    sheets,
-    hiddenHeader,
-    getCellStyle,
-  })
-  const workbook = {
-    SheetNames: sheets.map(({ name }) => name),
-    Sheets
-  }
-  const wbout = write(workbook, {
-    bookType: 'xlsx',
-    type: 'binary',
-  })
-  saveAs(
-    new Blob([s2ab(wbout)], { type: 'application/octet-stream' }),
-    filename
-  )
-}
+export default (args) => download(args, {
+  getCellStyle,
+  write
+})